CPU movement 95% complete
the cpu using a shortest path algorithm can randomly decided to move a normal or king checker to the closest red piece or pick a path to make it to the end to become king. Once it has decided it's move it can also actually make it happen and has checks in place to make sure it only lands in a free space. Now all it needs to be able to do is attack.