How to, win/lose, attack
Romoved player 2 option and added a link to a website to explain how to play. Added a win or lose message depending if black or red checkers hit zero. Added a basic way for CPU to attack a red checker. It only gets 5 attempts to "think" of a move before it gives up on attacking. So the difficulty is basically set to easy.